did fixes by PR comments
